refactor to adapt redux
