+++ client/src/QuestionAnswerNew/Components/addAnswer.js
@@ -1,10 +1,7 @@
 import React, { useState } from 'react';
-import { useSelector, useDispatch, connect } from 'react-redux';
-import allActions from '../../Actions';
 import axios from 'axios';
+
 const config = require('../../../../config.js');
-
-
 
 var AddAnswer = (props) => {
