+++ client/src/QuestionAnswerNew/Components/addQuestion.js
@@ -1,10 +1,8 @@
 import React, { useState } from 'react';
-import { useSelector, useDispatch, connect } from 'react-redux';
-import allActions from '../../Actions';
+import { useSelector } from 'react-redux';
 import axios from 'axios';
+
 const config = require('../../../../config.js');
-
-
 
 function AddQuestion(props) {
